requirements.txt
click==8.0.1
colorama==0.4.4
contextvars==2.4
Flask==2.0.1
Flask-MySQLdb==0.2.0
Flask-Reuploaded==1.1.0
Flask-Uploads==0.2.1
immutables==0.15
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
Pillow==8.2.0
Pillow-PIL==0.1.dev0
pytesseract==0.3.7
Werkzeug==2.0.0
